+++ mipengine/controller/algorithm_executor_node_data_objects.py
@@ -1,17 +1,4 @@
 from abc import ABC
-from typing import Any
-from typing import Dict
-from typing import List
-from typing import Union
-
-from mipengine.node_tasks_DTOs import NodeLiteralDTO
-from mipengine.node_tasks_DTOs import NodeSMPCDTO
-from mipengine.node_tasks_DTOs import NodeSMPCValueDTO
-from mipengine.node_tasks_DTOs import NodeTableDTO
-from mipengine.node_tasks_DTOs import NodeUDFDTO
-from mipengine.node_tasks_DTOs import TableSchema
-from mipengine.node_tasks_DTOs import UDFKeyArguments
-from mipengine.node_tasks_DTOs import UDFPosArguments
 
 
 class NodeData(ABC):
@@ -84,272 +71,3 @@ class NodeSMPCTables(NodeData):
         self.min_op = min_op
         self.max_op = max_op
         self.union_op = union_op
-
-
-class AlgoExecData(ABC):
-    """
-    AlgoExecData are representing one data object but could be located into
-    more than one node. For example the LocalNodesTable is treated as one table
-    from the algorithm executor but is located in multiple local nodes.
-    """
-
-    pass
-
-
-class LocalNodesData(AlgoExecData, ABC):
-    pass
-
-
-class GlobalNodeData(AlgoExecData, ABC):
-    pass
-
-
-# When _AlgorithmExecutionInterface::run_udf_on_local_nodes(..) is called, depending on
-# how many local nodes are participating in the current algorithm execution, several
-# database tables are created on all participating local nodes. Irrespectevely of the
-# number of local nodes participating, the number of tables created on each of these local
-# nodes will be the same.
-# Class _LocalNodeTable is the structure that represents the concept of these database
-# tables, created during the execution of a udf, in the algorithm execution layer. A key
-# concept is that a _LocalNodeTable stores 'pointers' to 'relevant' tables existing in
-# different local nodes accross the federation. By 'relevant' I mean tables that are
-# generated when triggering a udf execution accross several local nodes. By 'pointers'
-# I mean mapping between local nodes and table names and the aim is to hide the underline
-# complexity from the algorithm flow and exposing a single 'local node table' object that
-# stores in the background pointers to several tables in several local nodes.
-class LocalNodesTable(LocalNodesData):
-    def __init__(self, nodes_tables: Dict["LocalNode", NodeTable]):
-        self._nodes_tables = nodes_tables
-        self._validate_matching_table_names(list(self._nodes_tables.values()))
-
-    @property
-    def nodes_tables(self) -> Dict["LocalNode", NodeTable]:
-        return self._nodes_tables
-
-    # TODO this is redundant, either remove it or overload all node methods here?
-    def get_table_schema(self) -> TableSchema:
-        node = list(self.nodes_tables.keys())[0]
-        table = self.nodes_tables[node]
-        return node.get_table_schema(table)
-
-    def get_table_data(self) -> List[Union[int, float, str]]:
-        tables_data = []
-        for node, table_name in self.nodes_tables.items():
-            tables_data.append(node.get_table_data(table_name))
-        tables_data_flat = [table_data.columns for table_data in tables_data]
-        tables_data_flat = [
-            elem
-            for table in tables_data_flat
-            for column in table
-            for elem in column.data
-        ]
-        return tables_data_flat
-
-    def __repr__(self):
-        r = f"\n\tLocalNodeTable: {self.get_table_schema()}\n"
-        for node, table_name in self.nodes_tables.items():
-            r += f"\t{node=} {table_name=}\n"
-        return r
-
-    def _validate_matching_table_names(self, table_names: List[NodeTable]):
-        table_name_without_node_id = table_names[0].without_node_id()
-        for table_name in table_names:
-            if table_name.without_node_id() != table_name_without_node_id:
-                raise self.MismatchingTableNamesException(
-                    [table_name.full_table_name for table_name in table_names]
-                )
-
-    class MismatchingTableNamesException(Exception):
-        def __init__(self, table_names: List[str]):
-            message = f"Mismatched table names ->{table_names}"
-            super().__init__(message)
-            self.message = message
-
-
-class GlobalNodeTable(GlobalNodeData):
-    def __init__(self, node: "GlobalNode", table: NodeTable):
-        self._node = node
-        self._table = table
-
-    @property
-    def node(self) -> "GlobalNode":
-        return self._node
-
-    @property
-    def table(self) -> NodeTable:
-        return self._table
-
-    # TODO this is redundant, either remove it or overload all node methods here?
-    def get_table_schema(self) -> TableSchema:
-        return self._node.get_table_schema(self.table)
-
-    def get_table_data(self) -> List[List[Any]]:
-        table_data = [
-            column.data for column in self.node.get_table_data(self.table).columns
-        ]
-        return table_data
-
-    def __repr__(self):
-        r = f"\n\tGlobalNodeTable: \n\tschema={self.get_table_schema()}\n \t{self.table=}\n"
-        return r
-
-
-class LocalNodesSMPCTables(LocalNodesData):
-    template: LocalNodesTable
-    sum_op: LocalNodesTable
-    min_op: LocalNodesTable
-    max_op: LocalNodesTable
-    union_op: LocalNodesTable
-
-    def __init__(self, nodes_smpc_tables: Dict["LocalNode", NodeSMPCTables]):
-        template_nodes_tables = {}
-        sum_op_nodes_tables = {}
-        min_op_nodes_tables = {}
-        max_op_nodes_tables = {}
-        union_op_nodes_tables = {}
-        for node, node_smpc_tables in nodes_smpc_tables.items():
-            template_nodes_tables[node] = node_smpc_tables.template
-            sum_op_nodes_tables[node] = node_smpc_tables.sum_op
-            min_op_nodes_tables[node] = node_smpc_tables.min_op
-            max_op_nodes_tables[node] = node_smpc_tables.max_op
-            union_op_nodes_tables[node] = node_smpc_tables.union_op
-        self.template = LocalNodesTable(template_nodes_tables)
-        self.sum_op = create_local_nodes_table_from_nodes_tables(sum_op_nodes_tables)
-        self.min_op = create_local_nodes_table_from_nodes_tables(min_op_nodes_tables)
-        self.max_op = create_local_nodes_table_from_nodes_tables(max_op_nodes_tables)
-        self.union_op = create_local_nodes_table_from_nodes_tables(
-            union_op_nodes_tables
-        )
-
-
-class GlobalNodeSMPCTables(GlobalNodeData):
-    template: GlobalNodeTable
-    sum_op: GlobalNodeTable
-    min_op: GlobalNodeTable
-    max_op: GlobalNodeTable
-    union_op: GlobalNodeTable
-
-    def __init__(self, template, sum_op, min_op, max_op, union_op):
-        self.template = template
-        self.sum_op = sum_op
-        self.min_op = min_op
-        self.max_op = max_op
-        self.union_op = union_op
-
-
-def algoexec_udf_kwargs_to_node_udf_kwargs(
-    algoexec_kwargs: Dict[str, Any],
-    local_node: "LocalNode" = None,
-) -> UDFKeyArguments:
-    if not algoexec_kwargs:
-        return UDFKeyArguments(args={})
-
-    args = {}
-    for key, arg in algoexec_kwargs.items():
-        udf_argument = _algoexec_udf_arg_to_node_udf_arg(arg, local_node)
-        args[key] = udf_argument
-    return UDFKeyArguments(args=args)
-
-
-def algoexec_udf_posargs_to_node_udf_posargs(
-    algoexec_posargs: List[Any],
-    local_node: "LocalNode" = None,
-) -> UDFPosArguments:
-    if not algoexec_posargs:
-        return UDFPosArguments(args=[])
-
-    args = []
-    for arg in algoexec_posargs:
-        args.append(_algoexec_udf_arg_to_node_udf_arg(arg, local_node))
-    return UDFPosArguments(args=args)
-
-
-def _algoexec_udf_arg_to_node_udf_arg(
-    algoexec_arg: AlgoExecData, local_node: "LocalNode" = None
-) -> NodeUDFDTO:
-    """
-    Converts the algorithm executor run_udf input arguments, coming from the algorithm flow
-    to node udf pos/key arguments to be send to the NODE.
-
-    Parameters
-    ----------
-    algoexec_arg is the argument to be converted.
-    local_node is need only when the algoexec_arg is of LocalNodesTable, to know
-                which local table should be selected.
-
-    Returns
-    -------
-    a NodeUDFDTO
-    """
-    if isinstance(algoexec_arg, LocalNodesTable):
-        if not local_node:
-            raise ValueError(
-                "local_node parameter is required on LocalNodesTable conversion."
-            )
-        return NodeTableDTO(value=algoexec_arg.nodes_tables[local_node].full_table_name)
-    elif isinstance(algoexec_arg, GlobalNodeTable):
-        return NodeTableDTO(value=algoexec_arg.table.full_table_name)
-    elif isinstance(algoexec_arg, LocalNodesSMPCTables):
-        return NodeSMPCDTO(
-            value=NodeSMPCValueDTO(
-                template=algoexec_arg.template.nodes_tables[local_node].full_table_name,
-                sum_op_values=algoexec_arg.sum_op.nodes_tables[
-                    local_node
-                ].full_table_name,
-                min_op_values=algoexec_arg.min_op.nodes_tables[
-                    local_node
-                ].full_table_name,
-                max_op_values=algoexec_arg.max_op.nodes_tables[
-                    local_node
-                ].full_table_name,
-                union_op_values=algoexec_arg.union_op.nodes_tables[
-                    local_node
-                ].full_table_name,
-            )
-        )
-    elif isinstance(algoexec_arg, GlobalNodeSMPCTables):
-        return NodeSMPCDTO(
-            value=NodeSMPCValueDTO(
-                template=NodeTableDTO(
-                    value=algoexec_arg.template.table.full_table_name
-                ),
-                sum_op_values=create_node_table_dto_from_global_node_table(
-                    algoexec_arg.sum_op
-                ),
-                min_op_values=create_node_table_dto_from_global_node_table(
-                    algoexec_arg.min_op
-                ),
-                max_op_values=create_node_table_dto_from_global_node_table(
-                    algoexec_arg.max_op
-                ),
-                union_op_values=create_node_table_dto_from_global_node_table(
-                    algoexec_arg.union_op
-                ),
-            )
-        )
-    else:
-        return NodeLiteralDTO(value=algoexec_arg)
-
-
-def create_node_table_from_node_table_dto(node_table_dto: NodeTableDTO):
-    if not node_table_dto:
-        return None
-
-    return NodeTable(table_name=node_table_dto.value)
-
-
-def create_node_table_dto_from_global_node_table(node_table: GlobalNodeTable):
-    if not node_table:
-        return None
-
-    return NodeTableDTO(value=node_table.table.full_table_name)
-
-
-def create_local_nodes_table_from_nodes_tables(
-    nodes_tables: Dict["LocalNode", Union[NodeTable, None]]
-):
-    for table in nodes_tables.values():
-        if not table:
-            return None
-
-    return LocalNodesTable(nodes_tables)
